Added docs and rewrited examples for ydb (#17)
* Added docs and rewrited examples for ydb

+++ examples/authors/ydb/models.go
@@ -5,7 +5,7 @@
 package authors
 
 type Author struct {
-	ID   uint64  `json:"id"`
+	ID   int32   `json:"id"`
 	Name string  `json:"name"`
 	Bio  *string `json:"bio"`
 }
